Renamed services, errors and other did other requested changes
